Change score system
Scores are now computed on the basis of roles. Rather than being added up
directly they are now computed based on the assigned roles *after* the
evaluation.

Roles can also set a min and max score for an object, thus creating a lower
and upper bound for the achievable score.