Include ability to parser extra files
Normal text files e.g. README, TODOS, FAQ, LICENSE. For now this is to be
compatible with YARD, but I'm mot sure if I will keep this feature. How do
you measure a README or LICENSE file (except the fact that its existence is
a good thing and could give a thumbs-up).