Require return text for question mark methods
YARD assigns an implicit @return tag to methods ending in a question
mark. This and the fact, that a method ending in a question mark by
convention implies a Boolean return type, are the reasons why these
methods now have to include a decription of the returned value to
get the associated points.

Before this change, methods ending in a question mark would automatically
get some points because of their implicit return type and therefore be
graded 'C' even without any documentation.

After this, this does no longer happen. It does however not affect methods
previously ranked 'A', since the deducted points do not lead to a decrease
regarding the grade.