Update scoring rules to recognize @param in setter
Until now setters in Ruby were docced by their @return, which
was practical because we could handle `attr_accessor` the same
way.

Now, a setter needs either a @param (or a @return tag, for backwards
compatibility with the mentioned behaviour) to be documented.

Thanks to @paddor for bringing this up in inch-ci/inch_ci-web#109 :+1: