Prevent test suite from erring out partway through
Because the `CommandParser` had differing levels of responsibility
throughout its code paths, the test suite was erring out partway through
the test suite and not printing the results of running the whole suite.

By standardizing the `CommandParser.run` return value around a mock that
responds to `exit_status`, we leave the responsibility for exiting the
program to the `bin/inch` script instead of splitting it between the
script and the `CommandParser`. This change then allows the test suite
to continue past this test without issue.