Disable all warnings during test tasks
Currently, the implementation of attributes within the
`Inch::CLI::Command::Base` class causes warnings about uninitialized instance
variables, as it relies on a `define_method` that uses `instance_variable_get`
on uninitialized instance variables.

I couldn't find a simple way to redefine the accessor methods that I was sure
would maintain the intended behavior. This is largely because there are failing
tests in the test suite. The purpose of silencing all of these warnings was to
make it easier to understand what is happening in the test suite so we can track
down the failures and fix them.

As such, I think it is prudent to entirely disable warnings in the test suite
until we can get the test suite back to green.